refactor(DTO): DTO 계층 수정
DTO의 계층을 더 적합하도록 수정했습니다.
(Domain -> Application)
