+++ src/main/java/org/triumers/newsnippetback/Application/dto/AuthDTO.java
@@ -1,4 +1,4 @@
-package org.triumers.newsnippetback.domain.dto;
+package org.triumers.newsnippetback.Application.dto;
 
 import lombok.*;
 import org.triumers.newsnippetback.domain.aggregate.enums.Provider;
