+++ src/main/java/org/triumers/newsnippetback/Application/service/AuthService.java
@@ -4,9 +4,9 @@ import org.triumers.newsnippetback.common.exception.UserEmailDuplicateException;
 import org.triumers.newsnippetback.common.exception.UserNicknameDuplicateException;
 import org.triumers.newsnippetback.common.exception.WrongInputTypeException;
 import org.triumers.newsnippetback.common.exception.WrongPasswordException;
-import org.triumers.newsnippetback.domain.dto.AuthDTO;
-import org.triumers.newsnippetback.domain.dto.PasswordDTO;
-import org.triumers.newsnippetback.domain.dto.UserDTO;
+import org.triumers.newsnippetback.Application.dto.AuthDTO;
+import org.triumers.newsnippetback.Application.dto.PasswordDTO;
+import org.triumers.newsnippetback.Application.dto.UserDTO;
 
 public interface AuthService {
