+++ src/main/java/org/triumers/newsnippetback/Application/service/AuthServiceImpl.java
@@ -11,9 +11,9 @@ import org.triumers.newsnippetback.common.exception.WrongPasswordException;
 import org.triumers.newsnippetback.domain.aggregate.entity.User;
 import org.triumers.newsnippetback.domain.aggregate.enums.UserRole;
 import org.triumers.newsnippetback.domain.aggregate.enums.UserStatus;
-import org.triumers.newsnippetback.domain.dto.AuthDTO;
-import org.triumers.newsnippetback.domain.dto.PasswordDTO;
-import org.triumers.newsnippetback.domain.dto.UserDTO;
+import org.triumers.newsnippetback.Application.dto.AuthDTO;
+import org.triumers.newsnippetback.Application.dto.PasswordDTO;
+import org.triumers.newsnippetback.Application.dto.UserDTO;
 import org.triumers.newsnippetback.domain.repository.UserRepository;
 import org.triumers.newsnippetback.domain.service.ValidationAuthService;
