+++ src/main/java/org/triumers/newsnippetback/Application/service/ManageServiceImpl.java
@@ -7,8 +7,8 @@ import org.springframework.stereotype.Service;
 import org.triumers.newsnippetback.domain.aggregate.entity.Category;
 import org.triumers.newsnippetback.domain.aggregate.entity.CrawlingQuiz;
 import org.triumers.newsnippetback.domain.aggregate.entity.Quiz;
-import org.triumers.newsnippetback.domain.dto.CrawlingQuizDTO;
-import org.triumers.newsnippetback.domain.dto.QuizDTO;
+import org.triumers.newsnippetback.Application.dto.CrawlingQuizDTO;
+import org.triumers.newsnippetback.Application.dto.QuizDTO;
 import org.triumers.newsnippetback.domain.repository.CategoryRepository;
 import org.triumers.newsnippetback.domain.repository.CrawlingQuizRepository;
 import org.triumers.newsnippetback.domain.repository.QuizRepository;
