+++ src/main/java/org/triumers/newsnippetback/Application/service/QuizService.java
@@ -1,6 +1,6 @@
 package org.triumers.newsnippetback.Application.service;
 
-import org.triumers.newsnippetback.domain.dto.QuizDTO;
+import org.triumers.newsnippetback.Application.dto.QuizDTO;
 
 import java.time.LocalDate;
 import java.util.List;
