+++ src/main/java/org/triumers/newsnippetback/Application/service/SolvedService.java
@@ -3,7 +3,7 @@ package org.triumers.newsnippetback.Application.service;
 import org.triumers.newsnippetback.domain.aggregate.entity.Solved;
 import org.triumers.newsnippetback.domain.aggregate.vo.SolvedRequest;
 import org.triumers.newsnippetback.domain.aggregate.vo.SolvedResultRequest;
-import org.triumers.newsnippetback.domain.dto.SolvedDTO;
+import org.triumers.newsnippetback.Application.dto.SolvedDTO;
 
 import java.util.List;
