+++ src/test/java/org/triumers/newsnippetback/Application/service/ManageServiceTest.java
@@ -7,8 +7,8 @@ import org.junit.jupiter.params.provider.ValueSource;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.triumers.newsnippetback.domain.aggregate.entity.Quiz;
-import org.triumers.newsnippetback.domain.dto.CrawlingQuizDTO;
-import org.triumers.newsnippetback.domain.dto.QuizDTO;
+import org.triumers.newsnippetback.Application.dto.CrawlingQuizDTO;
+import org.triumers.newsnippetback.Application.dto.QuizDTO;
 
 import java.time.LocalDate;
 import java.util.List;
