+++ src/test/java/org/triumers/newsnippetback/Application/service/QuizServiceImplTest.java
@@ -3,11 +3,9 @@ package org.triumers.newsnippetback.Application.service;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.MethodSource;
-import org.junit.jupiter.params.provider.ValueSource;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
-import org.triumers.newsnippetback.domain.aggregate.vo.QuizRequest;
-import org.triumers.newsnippetback.domain.dto.QuizDTO;
+import org.triumers.newsnippetback.Application.dto.QuizDTO;
 
 import java.time.LocalDate;
 import java.util.List;
